More bugfixes

+++ HIPER/Helpers/Constants.cs
@@ -15,9 +15,9 @@ namespace HIPER.Helpers
         public const string HIPER_PEACH = "#ff7562";
         public const string CHALLENGE_GOLD = "#d4af37";
 
-        public const double POINTS_FOR_COMPLETED_GOAL = 10;
-        public const double POINTS_FOR_WON_CHALLENGE = 30;
-        public const double POINTS_FOR_ACCEPTED_CHALLENGE = 5;
+        public const int POINTS_FOR_COMPLETED_GOAL = 10;
+        public const int POINTS_FOR_WON_CHALLENGE = 30;
+        public const int POINTS_FOR_ACCEPTED_CHALLENGE = 5;
 
     }
 }
